move sample_blocks.hpp to core
