+++ silkworm/execution/grpc/server/endpoint/insertion_test.cpp
@@ -21,9 +21,9 @@
 #include <catch2/catch_test_macros.hpp>
 #include <intx/intx.hpp>
 
+#include <silkworm/core/test_util/sample_blocks.hpp>
 #include <silkworm/infra/test_util/fixture.hpp>
 #include <silkworm/interfaces/execution/execution.pb.h>
-#include <silkworm/node/test_util/sample_blocks.hpp>
 
 #include "../../test_util/sample_protos.hpp"
